+++ src/components/Fields/Section.tsx
@@ -2,13 +2,14 @@ import type { ObjectFieldTemplateProps } from '@rjsf/core';
 import DropZone from '../EditorFieldTemplate/DropZone';
 import { useVerna } from '../../providers/VernaProvider';
 
-export default function Section(props: ObjectFieldTemplateProps) {
+export default function Section({ title, idSchema, properties }: ObjectFieldTemplateProps) {
   const { isEditor } = useVerna();
 
   return (
     <fieldset>
-      {isEditor && <DropZone id={props.idSchema['$id']} />}
-      {props.properties.map((element) => element.content)}
+      <legend>{title}</legend>
+      {isEditor && <DropZone id={idSchema['$id']} />}
+      {properties.map((element) => element.content)}
     </fieldset>
   );
 }