+++ src/types/rjsf.tsx
@@ -1,10 +1,13 @@
 import { JSONSchema7 } from 'json-schema';
 
-type VernaJSONSchemaType = Omit<JSONSchema7, 'properties' | 'items'> & {
-  properties?: {
-    [key: string]: VernaJSONSchemaType;
-  };
+export type RJSFPropertiesType = {
+  [key: string]: VernaJSONSchemaType;
+};
+
+type VernaJSONSchemaType = Omit<JSONSchema7, 'properties' | 'items' | 'enum'> & {
+  properties?: RJSFPropertiesType;
   items?: VernaJSONSchemaType | VernaJSONSchemaType[];
+  enum?: string[];
 };
 
 export default VernaJSONSchemaType;