+++ src/types/translations.tsx
@@ -1,0 +1,5 @@
+import type { ResolvedIntlConfig } from 'react-intl';
+
+export interface TranslationType {
+  [locale: string]: ResolvedIntlConfig['messages'];
+}